requirements.txt
numpy==1.20.3
scikit-learn
joblib==1.1.0
streamlit==1.9.0
openapi-codec==1.3.2
pandas==1.4.2
Pillow==9.1.1